refactored the react components
