+++ packages/reporters/src/components/ThreadItem.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+import { Box, Text } from "ink";
+
+export function ThreadItem(props: { targetId: string }) {
+  const { targetId } = props;
+  return <Box>{targetId ? <Text color="whiteBright">[ {targetId} ]</Text> : <Text color="gray">[ IDLE ]</Text>}</Box>;
+}
